+++ moodle/Tests/Sniffs/Commenting/InlineCommentSniffTest.php
@@ -29,22 +29,46 @@ use MoodleHQ\MoodleCS\moodle\Tests\MoodleCSBaseTestCase;
  */
 class InlineCommentSniffTest extends MoodleCSBaseTestCase
 {
-    public function testCommentBeforeAttribute(): void {
-        // Define the standard, sniff and fixture to use.
+    /**
+     * @dataProvider commentsProvider
+     */
+    public function testComments(
+        string $fixture,
+        ?string $fixtureFilename,
+        array $errors,
+        array $warnings
+    ): void {
         $this->setStandard('moodle');
         $this->setSniff('moodle.Commenting.InlineComment');
-        $this->setFixture(__DIR__ . '/../../fixtures/Commenting/InlineCommentAttributeAfter.php');
-
-        // Define expected results (errors and warnings). Format, array of:
-        // - line => number of problems,  or
-        // - line => array of contents for message / source problem matching.
-        // - line => string of contents for message / source problem matching (only 1).
-        $errors = [];
-        $warnings = [];
-        $this->setErrors($errors);
+        $this->setFixture(sprintf("%s/fixtures/InlineComment/%s.php", __DIR__, $fixture), $fixtureFilename);
         $this->setWarnings($warnings);
+        $this->setErrors($errors);
+        $this->setComponentMapping([
+            'local_codechecker' => dirname(__DIR__),
+        ]);
 
-        // Let's do all the hard work!
         $this->verifyCsResults();
+    }
+
+    public static function commentsProvider(): \Generator {
+        yield '' => [
+            'fixture' => 'attributes',
+            'fixtureFilename' => null,
+            'errors' => [],
+            'warnings' => [],
+        ];
+        yield 'Closing punctuation behaves correctly' => [
+            'fixture' => 'punctuation',
+            'fixtureFilename' => null,
+            'errors' => [
+            ],
+            'warnings' => [
+                38 => 'Inline comments must end in ',
+                54 => 'Inline comments must end in ',
+                61 => 'Inline comments must end in ',
+                63 => 'Inline comments must end in ',
+                65 => 'Inline comments must end in ',
+            ],
+        ];
     }
 }